requirements.txt
boto3==1.9.38
certifi==2018.8.24
chardet==3.0.4
dateparser==0.7.0
Django==2.1.5
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-imagekit==4.0.2
django-markdownx==2.0.28
django-modeltranslation==0.13b1
django-settings-export==1.2.1
django-storages==1.7.1
idna==2.7
Markdown==3.0.1
pilkit==2.0
Pillow==5.2.0
psycopg2-binary==2.7.5
pyinaturalist==0.6.0
python-dateutil==2.7.3
pytz==2018.5
regex==2018.8.29
requests==2.20.0
six==1.11.0
sqlparse==0.2.4
tzlocal==1.5.1
urllib3==1.23
